Add files via upload
Il progetto non ha scopro di lucro e/o pubblicitario, è solo un'esercitazione a tema per un corso di front-end developer. La scelta dell'artista in questione è dettata esclusivamente dal gusto personale della corsista.

The project is not for profit and/or advertising, it is just a themed exercise for a front-end developer course. The choice of the artist in question is dictated exclusively by the personal taste of the student.